src/backend/requirements.txt
asgiref==3.8.1
Django==5.1.1
django-cors-headers==4.6.0
django-debug-toolbar==4.4.6
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
Markdown==3.7
pillow==10.4.0
PyJWT==2.9.0
sqlparse==0.5.1
